requirements.txt
streamlit
langchain
langchain-core
langchain-openai
openai
yt-dlp
requests
python-dotenv